+++ apps/golang/commons/monitoring/newrelic/entity/guid.go
